integration_tests/test-requirements.txt
docker-compose
pyhamcrest
nose
python-consul==0.4.2
